登录接入接口，接口附带token

+++ src/store/modules/user.js
@@ -2,7 +2,7 @@
  * @Author: eamiear
  * @Date: 2019-02-06 18:37:25
  * @Last Modified by: eamiear
- * @Last Modified time: 2019-12-04 16:49:05
+ * @Last Modified time: 2019-12-09 15:53:18
  */
 
 import {
@@ -14,6 +14,7 @@ import {
   SET_USER_INFO,
   SET_PWD
 } from '../mutation-types'
+import SystemAPI from '@/api/system'
 import Storage, {cacher} from '@/common/cache'
 import md5 from 'md5'
 
@@ -21,11 +22,11 @@ const user = {
   state: {
     user: '',
     token: Storage.getToken(),
-    name: cacher.setStrategy('localStorage').get('name'),
+    name: cacher.setStrategy('localStorage').get('vms.name'),
     avatar: '',
     introduction: '',
     userInfo: null,
-    pwd: cacher.setStrategy('localStorage').get('pk')
+    pwd: cacher.setStrategy('localStorage').get('vms.pk')
   },
   mutations: {
     [SET_TOKEN] (state, token) {
@@ -52,29 +53,44 @@ const user = {
   },
   actions: {
     loginByAccount ({ commit }, userInfo) {
-      return new Promise(resolve => {
+      return new Promise((resolve, reject) => {
         const password = md5(btoa(userInfo.password) + userInfo.password)
-        cacher.setStrategy('localStorage').set('pk', password)
-        cacher.setStrategy('localStorage').set('name', userInfo.account.trim())
-        Storage.setToken(password)
-        commit('SET_TOKEN', password)
-        commit('SET_NAME', userInfo.account.trim())
-        commit('SET_PWD', password)
-        resolve(password)
+        SystemAPI.login(userInfo.account.trim(), password).then(data => {
+          if (data) {
+            const token = data.access_token
+            Storage.setToken(token)
+            cacher.setStrategy('localStorage').set('vms.pk', password)
+            cacher.setStrategy('localStorage').set('vms.name', userInfo.account.trim())
+            commit('SET_TOKEN', token)
+            commit('SET_NAME', userInfo.account.trim())
+            commit('SET_USER_INFO', data)
+            commit('SET_PWD', password)
+          }
+          resolve(data)
+        }).catch(error => {
+          reject(error)
+        })
       })
     },
     // 登出
     logOut ({ commit, state }) {
-      return new Promise(resolve => {
-        commit('SET_TOKEN', '')
-        commit('SET_USER_INFO', null)
-        commit('SET_NAME', '')
-        commit('SET_PWD', '')
-        cacher.remove('permission')
-        Storage.removeToken()
-        cacher.setStrategy('localStorage').remove('pk')
-        cacher.setStrategy('localStorage').remove('name')
-        resolve()
+      return new Promise((resolve, reject) => {
+        SystemAPI.logout(state.token).then(res => {
+          if (res.message.includes('success')) {
+            commit('SET_TOKEN', '')
+            commit('SET_USER_INFO', null)
+            commit('SET_NAME', '')
+            commit('SET_PWD', '')
+            Storage.removeToken()
+            cacher.setStrategy('localStorage').remove('vms.pk')
+            cacher.setStrategy('localStorage').remove('vms.name')
+            resolve()
+          } else {
+            reject(res.message)
+          }
+        }).catch(error => {
+          reject(error)
+        })
       })
     }
   }
